Creates add and edit task modals
Uses new create modal function to add and edit task
modals. Creates new auto-incrementing task and project
ids for keeping track of these objects when created
and deleted. SingleTaskDialogComponent is now
unneeded and removed. Currently trying to change the
project of a task after it has been created breaks
and this still needs to be fixed